CandyWrappers: Import our new wallpapers
Shoutout to Aaron Wieneke and Jason Frawley, our core graphics guys,
as well as the people who have made contributions to our graphics over
the years and preceding months! We appreciate it.

In addition, thanks to Mathew Engberg, Ozz Allsop, and Joseph Powell
for the contributions to graphics early on.